requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
contourpy==1.3.3
coverage==7.13.5
cycler==0.12.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
fonttools==4.62.1
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pluggy==1.6.0
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2026.1
urllib3==2.6.3
